Extract account and general settings sections
